User_MicroService_Group3/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
gunicorn==21.2.0
kafka-python==2.0.2
pyodbc==5.1.0
blinker==1.7.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
Werkzeug==3.0.1
boto3==1.34.71
colorama==0.4.6
MarkupSafe==2.1.3
requests==2.31.0